Add 'Improve this page' button
Out website is public and accepts user contributions, we should provide
an easy link for users who want to make a quick contribution, like
fixing a typo, to find the relevant repo and file within the website
repo to edit.

Also updating the github-pages gem so that it builds with the latest
ruby.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>